feat: isolated the types creation logic
