+++ lib/src/functions/api/database.js
@@ -3,6 +3,11 @@ const { promisify } = require('util')
 const exec = promisify(require('child_process').exec)
 const writeFile = require('../../utils/writeFile')
 
+/**
+ * @param {Object} args
+ * @param {Boolean|undefined} args.mongo
+ * @param {String} args.projectName
+ */
 module.exports = async ({ mongo = true, projectName }) => {
   const createFoldersCommand = `mkdir ${projectName}/src/database \
   ${projectName}/src/database/mongo \
