+++ lib/src/functions/api/index.js
@@ -1,10 +1,12 @@
 const os = require('os')
 const util = require('util')
 const exec = util.promisify(require('child_process').exec)
-const writeFile = require('../../utils/writeFile')
+
 const database = require('./database')
 const schemas = require('./schemas')
 const services = require('./services')
+const types = require('./types')
+const writeFile = require('../../utils/writeFile')
 
 /*
  * Express api:
@@ -116,29 +118,6 @@ module.exports = async ({
   mongo = true
 }) => {
   const data = {
-    '@types': {
-      index: {
-        content: `/* eslint-disable no-var */
-declare global {}
-
-export {}
-`,
-        file: `${projectName}/src/@types/index.d.ts`
-      }
-    },
-    '@types/models': {
-      user: {
-        content: `interface UserDBO {
-  id: string
-  name: string
-  lastName: string
-  createdAt: Date
-  updatedAt: Date
-}
-`,
-        file: `${projectName}/src/@types/models/user.d.ts`
-      }
-    },
     network: {
       index: {
         content: `export * from './routes'
@@ -203,33 +182,6 @@ Server.start()
     }
   }
   const expressData = {
-    '@types/custom': {
-      request: {
-        content: `type ExpressRequest = import('express').Request
-
-interface CustomRequest extends ExpressRequest {
-  body: {
-    args?: import('schemas').UserDTO
-  }
-  // We can add custom headers via intersection, remember that for some reason
-  // headers must be in Snake-Pascal-Case
-  headers: import('http').IncomingHttpHeaders & {
-    'Custom-Header'?: string
-  }
-}
-`,
-        file: `${projectName}/src/@types/custom/request.d.ts`
-      },
-      response: {
-        content: `type ExpressResponse = import('express').Response
-
-interface CustomResponse extends ExpressResponse {
-  newValue?: string
-}
-`,
-        file: `${projectName}/src/@types/custom/response.d.ts`
-      }
-    },
     network: {
       response: {
         content: `interface ResponseProps {
@@ -1491,14 +1443,11 @@ export { validatorCompiler }
   }
 
   const expressFolders = `${projectName}/src/utils \
-${projectName}/src/@types/custom \
 ${projectName}/src/network/routes/utils`
 
   const fastifyFolders = `${projectName}/src/network/utils`
 
   const createFoldersCommands = `mkdir ${projectName}/src \
-${projectName}/src/@types \
-${projectName}/src/@types/models \
 ${projectName}/src/network \
 ${projectName}/src/network/routes \
 ${fastify ? `${fastifyFolders}` : `${expressFolders}`}
@@ -1508,6 +1457,11 @@ ${fastify ? `${fastifyFolders}` : `${expressFolders}`}
     await exec(createFoldersCommands.replaceAll('/', '\\'))
   else await exec(createFoldersCommands)
 
+  // /@types
+  types({
+    express: !fastify,
+    projectName
+  })
   // /database
   database({
     mongo,
@@ -1517,15 +1471,6 @@ ${fastify ? `${fastifyFolders}` : `${expressFolders}`}
   schemas({ projectName })
   // /services
   services({ projectName })
-
-  // /@types
-  await writeFile(data['@types'].index.file, data['@types'].index.content)
-
-  // /@types/models
-  await writeFile(
-    data['@types/models'].user.file,
-    data['@types/models'].user.content
-  )
 
   // /network
   await writeFile(data.network.index.file, data.network.index.content)
@@ -1578,16 +1523,6 @@ ${fastify ? `${fastifyFolders}` : `${expressFolders}`}
       fastifyData['network/utils'].index.content
     )
   } else {
-    // /@types/custom
-    await writeFile(
-      expressData['@types/custom'].request.file,
-      expressData['@types/custom'].request.content
-    )
-    await writeFile(
-      expressData['@types/custom'].response.file,
-      expressData['@types/custom'].response.content
-    )
-
     // /network
     await writeFile(
       expressData.network.response.file,
