+++ lib/src/functions/api/schemas.js
@@ -3,6 +3,10 @@ const { promisify } = require('util')
 const exec = promisify(require('child_process').exec)
 const writeFile = require('../../utils/writeFile')
 
+/**
+ * @param {Object} args
+ * @param {String} args.projectName
+ */
 module.exports = async ({ projectName }) => {
   const createFoldersCommand = `mkdir ${projectName}/src/schemas`
